% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city-data.R
\docType{data}
\name{city}
\alias{city}
\title{City Code Data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1046 rows and 5 columns.}
\source{
\href{https://ec.europa.eu/eurostat/web/cities/spatial-units}{Spatial Units}
}
\usage{
data(city)
}
\description{
Data containing Eurostat functional urban area (fua) and city names and codes
along with the country each fua belongs to.
}
\keyword{datasets}
