% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags-data.R
\docType{data}
\name{flags}
\alias{flags}
\title{Eurostat Flags and Flag Descriptions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 2 columns.}
\source{
\href{https://ec.europa.eu/eurostat/statistics-explained/index.php?title=Tutorial:Symbols_and_abbreviations#Statistical_symbols.2C_abbreviations_and_units_of_measurement}{Statistical symbols, abbreviations and units of measurement}
}
\usage{
data(flags)
}
\description{
Data pertaining to the flags associated with some of the values in Eurostat.
}
\keyword{datasets}
