% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fua-data.R
\docType{data}
\name{fua}
\alias{fua}
\title{Functional Urban Areas (FUA) Data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1046 rows and 4 columns.}
\source{
\href{https://ec.europa.eu/eurostat/web/cities/spatial-units}{Spatial Units}
}
\usage{
data(fua)
}
\description{
Data containing Eurostat functional urban area (fua) names and codes along
with the country each fua belongs to.
}
\keyword{datasets}
