% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_stats.R
\name{get_city_stats}
\alias{get_city_stats}
\title{Download City Statistics datasets.}
\usage{
get_city_stats(dataset, cities, language = "en", props = FALSE)
}
\arguments{
\item{dataset}{The dataset code pertaining to the variable of interest.}

\item{cities}{The functional urban area code or a vector of codes for the
functional urban areas of interest.}

\item{language}{The language of the data.}

\item{props}{An option to include proportional data along with absolute
values.}
}
\value{
A tibble of with the attributes `urb_code`, `city`, `year`, `value`,
  `var_code` and `variable`.
}
\description{
Description from eurostat (see link for reference).
The Functional Urban Area consists of a city and its commuting zone. (This
was formerly known as larger urban zone (LUZ)).
}
\details{
Download data pertaining to Functional Urban Areas, cities and greater cities
from eurostat.
}
\examples{
\dontrun{
library(reurostat)

brus_antw <- get_city_stat(dataset = "urb_lpop1", cities = c("BE002L2", "BE001L2"))
}
}
\seealso{
\url{https://ec.europa.eu/eurostat/web/cities/spatial-units}
}
