% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_urban_areas.R
\name{get_fua}
\alias{get_fua}
\title{Download Functional Urban Areas datasets.}
\usage{
get_fua(dataset, cities, language = "en", props = FALSE)
}
\arguments{
\item{dataset}{The dataset code pertaining to the variable of interest.}

\item{cities}{The functional urban area code or a vector of codes for the
functional urban areas of interest.}

\item{language}{The language of the data.}

\item{props}{An option to include proportional data along with absolute
values.}
}
\value{
A tibble of with the attributes `fua_code`, `fua`, `year`, `value`,
  `code` and `variable`.
}
\description{
Description from eurostat (see link for reference).
The Functional Urban Area consists of a city and its commuting zone. (This
was formerly known as larger urban zone (LUZ)).
}
\details{
Download data pertaining to Functional Urban Areas from eurostat. Cities
here refers to the major ubran area in which the Functional Urban Area
belongs to.
}
\examples{
\dontrun{
library(reurostat)

brus_antw <- get_fua(dataset = "urb_lpop1", cities = c("BE002L2", "BE001L2"))
}
}
\seealso{
\url{https://ec.europa.eu/eurostat/web/cities/spatial-units}
}
