% Generated by roxygen2 (4.0.1): do not edit by hand
\name{helloreurostat}
\alias{helloreurostat}
\title{Print hello}
\usage{
helloreurostat(fname, lname)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function prints hello
}
\examples{
helloreurostat(fname="Albrecht",lname="Gradmann")
helloreurostat(fname="Stefan",lname="Luedtke")
}

