% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estat_index}
\alias{estat_index}
\title{Loads the index of available datasets from Eurostat.}
\usage{
estat_index(lang = c("en", "de", "fr"),
  url = "http://ec.europa.eu/eurostat/SDMX/diss-web/rest/dataflow/ESTAT/all/latest")
}
\arguments{
\item{lang}{language used to display name of dataflow. \code{en} for english (default),
\code{de} for german or \code{fr} for french.}

\item{url}{address of the dataflow list (REST)}
}
\value{
dataframe with two variables ID and name
}
\description{
This function downloads the current index of available datasets from Eurostat
using the websites REST service.
}

